refactor: Restructure product-related packages based on screen flow (#40)
This commit refactors the project structure by moving several screens and components to more specific feature packages.

- Renamed package of CreatePostScreen, CreatePostViewModel, and CategoryTagDrawer to feature.createpost to better reflect feature scope.
- Product Detail related screens and components have been moved under a new `feature.productdetail` package.
