+++ app/src/main/java/com/example/rentit/feature/createpost/component/TagButton.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.feature.product.component
+package com.example.rentit.feature.createpost.component
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
