+++ app/src/main/java/com/example/rentit/feature/productdetail/rentalhistory/UsageDetailBottomDrawer.kt
@@ -1,19 +1,14 @@
-package com.example.rentit.feature.product
+package com.example.rentit.feature.productdetail.rentalhistory
 
 import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.example.rentit.feature.product.component.calendar.ReadOnlyCalender
+import com.example.rentit.feature.productdetail.rentalhistory.components.ReadOnlyCalender
 import com.example.rentit.common.theme.RentItTheme
-import com.example.rentit.feature.product.component.UsageListItem
 import java.time.YearMonth
 
 @RequiresApi(Build.VERSION_CODES.O)
